snowpark/requirements.txt
snowflake-snowpark-python==1.10.0
